Add knight integration test
- split fixtures to different file
- added valid piece moves test
